+++ tests/SeleniumTests/__init__.py
@@ -18,6 +18,7 @@ SELENIUM_GRID_USERNAME = os.environ.get('SELENIUM_GRID_USERNAME', '')
 SELENIUM_GRID_PASSWORD = os.environ.get('SELENIUM_GRID_PASSWORD', '')
 SELENIUM_GRID_TEST_HEADLESS = os.environ.get('SELENIUM_GRID_TEST_HEADLESS', 'false').lower() == 'true'
 WEB_DRIVER_WAIT_TIMEOUT = int(os.environ.get('WEB_DRIVER_WAIT_TIMEOUT', 60))
+TEST_PARALLEL_HARDENING = os.environ.get('TEST_PARALLEL_HARDENING', 'false').lower() == 'true'
 
 if SELENIUM_GRID_USERNAME and SELENIUM_GRID_PASSWORD:
     SELENIUM_GRID_HOST = f"{SELENIUM_GRID_USERNAME}:{SELENIUM_GRID_PASSWORD}@{SELENIUM_GRID_HOST}"
@@ -87,6 +88,7 @@ class SeleniumGenericTests(unittest.TestCase):
         file_link = wait.until(
             EC.element_to_be_clickable((By.LINK_TEXT, file_name))
         )
+        driver.execute_script("arguments[0].scrollIntoView();", file_link)
         file_link.click()
         wait.until(
             lambda d: str(d.get_downloadable_files()[0]).endswith(file_name)
@@ -185,6 +187,7 @@ class Autoscaling():
                     start_times[test] = time.time()
                     futures.append(executor.submit(test))
                     tests.append(test)
+            print(f"Number of tests were added to worker: {len(tests)}")
             failed_tests = []
             for future, test in zip(concurrent.futures.as_completed(futures), tests):
                 try:
@@ -215,9 +218,19 @@ class Autoscaling():
 class DeploymentAutoscalingTests(unittest.TestCase):
     def test_parallel_autoscaling(self):
         runner = Autoscaling()
-        runner.run([FirefoxTests, EdgeTests, ChromeTests])
+        if not TEST_PARALLEL_HARDENING:
+            runner.run([FirefoxTests, EdgeTests, ChromeTests])
+        else:
+            runner.run([FirefoxTests, EdgeTests, ChromeTests, FirefoxTests, EdgeTests, ChromeTests, FirefoxTests, EdgeTests, ChromeTests,
+                        FirefoxTests, EdgeTests, ChromeTests, FirefoxTests, EdgeTests, ChromeTests, FirefoxTests, EdgeTests, ChromeTests,
+                        FirefoxTests, EdgeTests, ChromeTests, FirefoxTests, EdgeTests, ChromeTests, FirefoxTests, EdgeTests, ChromeTests,])
 
 class JobAutoscalingTests(unittest.TestCase):
     def test_parallel_autoscaling(self):
         runner = Autoscaling()
-        runner.run([FirefoxTests, EdgeTests, ChromeTests])
+        if not TEST_PARALLEL_HARDENING:
+            runner.run([FirefoxTests, EdgeTests, ChromeTests])
+        else:
+            runner.run([FirefoxTests, EdgeTests, ChromeTests, FirefoxTests, EdgeTests, ChromeTests, FirefoxTests, EdgeTests, ChromeTests,
+                        FirefoxTests, EdgeTests, ChromeTests, FirefoxTests, EdgeTests, ChromeTests, FirefoxTests, EdgeTests, ChromeTests,
+                        FirefoxTests, EdgeTests, ChromeTests, FirefoxTests, EdgeTests, ChromeTests, FirefoxTests, EdgeTests, ChromeTests,])