+++ tests/charts/bootstrap.sh
@@ -12,20 +12,39 @@ python -m pip install pyyaml==6.0.1 \
                       | grep -v 'Requirement already satisfied'
 
 cd ..
-helm template dummy --values tests/charts/templates/render/dummy.yaml \
-  --set-file 'nodeConfigMap.extraScripts.nodePreStop\.sh=tests/charts/templates/render/dummy_external.sh' \
-  --set-file 'recorderConfigMap.extraScripts.video\.sh=tests/charts/templates/render/dummy_external.sh' \
-  --set-file 'recorderConfigMap.extraScripts.video_graphQLQuery\.sh=tests/charts/templates/render/dummy_external.sh' \
-  --set-file 'recorderConfigMap.extraScripts.newInsertScript\.sh=tests/charts/templates/render/dummy_external.sh' \
-  --set-file 'uploaderConfigMap.extraScripts.upload\.sh=tests/charts/templates/render/dummy_external.sh' \
-  --set-file 'uploaderConfigMap.secretFiles.upload\.conf=tests/charts/templates/render/dummy_external.sh' \
+
+RELEASE_NAME="selenium"
+
+helm template ${RELEASE_NAME} --values tests/charts/templates/render/dummy.yaml \
+  --set-file 'nodeConfigMap.extraScripts.setFromCommand\.sh=tests/charts/templates/render/dummy_external.sh' \
+  --set-file 'recorderConfigMap.extraScripts.setFromCommand\.sh=tests/charts/templates/render/dummy_external.sh' \
+  --set-file 'uploaderConfigMap.extraScripts.setFromCommand\.sh=tests/charts/templates/render/dummy_external.sh' \
   charts/selenium-grid > ./tests/tests/dummy_template_manifests.yaml
 
-python tests/charts/templates/test.py "./tests/tests/dummy_template_manifests.yaml" dummy
-ret_code=$?
+python tests/charts/templates/test.py "./tests/tests/dummy_template_manifests.yaml" ${RELEASE_NAME}
+if [ $? -ne 0 ]; then
+  echo "Failed to validate the chart"
+  exit 1
+fi
+
+rm -rf tests/charts/umbrella-charts/Chart.lock tests/charts/umbrella-charts/charts
+helm dependency update tests/charts/umbrella-charts
+helm dependency build tests/charts/umbrella-charts
+
+RELEASE_NAME="test"
+
+helm template ${RELEASE_NAME} --values tests/charts/templates/render/dummy_solution.yaml \
+  --set-file 'selenium-grid.nodeConfigMap.extraScripts.setFromCommand\.sh=tests/charts/templates/render/dummy_external.sh' \
+  --set-file 'selenium-grid.recorderConfigMap.extraScripts.setFromCommand\.sh=tests/charts/templates/render/dummy_external.sh' \
+  --set-file 'selenium-grid.uploaderConfigMap.extraScripts.setFromCommand\.sh=tests/charts/templates/render/dummy_external.sh' \
+  tests/charts/umbrella-charts > ./tests/tests/dummy_solution_template_manifests.yaml
+
+python tests/charts/templates/test.py "./tests/tests/dummy_solution_template_manifests.yaml" ${RELEASE_NAME}
+if [ $? -ne 0 ]; then
+  echo "Failed to validate the umbrella chart"
+  exit 1
+fi
 
 if [ "${CI:-false}" = "false" ]; then
   deactivate
 fi
-
-exit $ret_code