+++ tests/charts/make/chart_test.sh
@@ -16,29 +16,39 @@ SELENIUM_GRID_HOST=${SELENIUM_GRID_HOST:-"localhost"}
 SELENIUM_GRID_PORT=${SELENIUM_GRID_PORT:-"80"}
 MATRIX_BROWSER=${1:-"NodeChrome"}
 SELENIUM_GRID_AUTOSCALING=${2:-"true"}
-SELENIUM_GRID_AUTOSCALING_MIN_REPLICA=${3:-"0"}
+SELENIUM_GRID_AUTOSCALING_MIN_REPLICA=${SELENIUM_GRID_AUTOSCALING_MIN_REPLICA:-"0"}
 WAIT_TIMEOUT=${WAIT_TIMEOUT:-"90s"}
 HUB_CHECKS_INTERVAL=${HUB_CHECKS_INTERVAL:-45}
 HUB_CHECKS_MAX_ATTEMPTS=${HUB_CHECKS_MAX_ATTEMPTS:-6}
 WEB_DRIVER_WAIT_TIMEOUT=${WEB_DRIVER_WAIT_TIMEOUT:-120}
 AUTOSCALING_POLL_INTERVAL=${AUTOSCALING_POLL_INTERVAL:-20}
-SKIP_CLEANUP=${SKIP_CLEANUP:-"false"} # For debugging purposes, retain the cluster after the test run
+SKIP_CLEANUP=${SKIP_CLEANUP:-"true"} # For debugging purposes, retain the cluster after the test run
 CHART_CERT_PATH=${CHART_CERT_PATH:-"${CHART_PATH}/certs/selenium.pem"}
 SSL_CERT_DIR=${SSL_CERT_DIR:-"/etc/ssl/certs"}
 VIDEO_TAG=${VIDEO_TAG:-"latest"}
 CHART_ENABLE_TRACING=${CHART_ENABLE_TRACING:-"false"}
 CHART_FULL_DISTRIBUTED_MODE=${CHART_FULL_DISTRIBUTED_MODE:-"false"}
-HOSTNAME_ADDRESS=${HOSTNAME_ADDRESS:-"selenium-grid.local"}
+HOSTNAME_ADDRESS=${HOSTNAME_ADDRESS:-"selenium-grid.prod"}
 CHART_ENABLE_INGRESS_HOSTNAME=${CHART_ENABLE_INGRESS_HOSTNAME:-"false"}
 CHART_ENABLE_BASIC_AUTH=${CHART_ENABLE_BASIC_AUTH:-"false"}
 BASIC_AUTH_USERNAME=${BASIC_AUTH_USERNAME:-"sysAdminUser"}
 BASIC_AUTH_PASSWORD=${BASIC_AUTH_PASSWORD:-"myStrongPassword"}
+LOG_LEVEL=${LOG_LEVEL:-"INFO"}
+TEST_EXISTING_KEDA=${TEST_EXISTING_KEDA:-"true"}
 
 cleanup() {
+  # Get the list of pods
+  pods=$(kubectl get pods -n ${SELENIUM_NAMESPACE} -o jsonpath='{.items[*].metadata.name}')
+  # Iterate over the pods and print their logs
+  for pod in $pods; do
+    echo "Logs for pod $pod"
+    kubectl logs -n ${SELENIUM_NAMESPACE} $pod > tests/tests/pod_logs_${pod}.txt
+  done
   if [ "${SKIP_CLEANUP}" = "false" ]; then
     echo "Clean up chart release and namespace"
-    helm delete ${RELEASE_NAME} --namespace ${SELENIUM_NAMESPACE}
-    kubectl delete namespace ${SELENIUM_NAMESPACE}
+    helm delete ${RELEASE_NAME} --namespace ${SELENIUM_NAMESPACE} --wait
+    kubectl patch ns ${SELENIUM_NAMESPACE} -p '{"metadata":{"finalizers":null}}'
+    kubectl delete namespace ${SELENIUM_NAMESPACE} --wait=false
   fi
 }
 
@@ -56,6 +66,7 @@ on_failure() {
 # Trap ERR signal and call on_failure function
 trap 'on_failure' ERR EXIT
 
+rm -rf tests/tests/*
 touch tests/tests/describe_all_resources_${MATRIX_BROWSER}.txt
 
 if [ -f .env ]
@@ -65,6 +76,11 @@ else
     export UPLOAD_ENABLED=false
 fi
 export RELEASE_NAME=${RELEASE_NAME}
+if [ "${RELEASE_NAME}" = "selenium" ]; then
+  export SELENIUM_TLS_SECRET_NAME="selenium-tls-secret"
+else
+  export SELENIUM_TLS_SECRET_NAME="${RELEASE_NAME}-selenium-tls-secret"
+fi
 RECORDER_VALUES_FILE=${TEST_VALUES_PATH}/base-recorder-values.yaml
 envsubst < ${RECORDER_VALUES_FILE} > ./tests/tests/base-recorder-values.yaml
 RECORDER_VALUES_FILE=./tests/tests/base-recorder-values.yaml
@@ -77,7 +93,26 @@ HELM_COMMAND_SET_IMAGES=" \
 --set autoscaling.scaledOptions.pollingInterval=${AUTOSCALING_POLL_INTERVAL} \
 --set tracing.enabled=${CHART_ENABLE_TRACING} \
 --set isolateComponents=${CHART_FULL_DISTRIBUTED_MODE} \
+--set global.seleniumGrid.logLevel=${LOG_LEVEL} \
 "
+
+if [ "${TEST_EXISTING_KEDA}" = "true" ]; then
+  HELM_COMMAND_SET_IMAGES="${HELM_COMMAND_SET_IMAGES} \
+  --set autoscaling.enabled=false \
+  --set autoscaling.enableWithExistingKEDA=true \
+  "
+else
+  HELM_COMMAND_SET_IMAGES="${HELM_COMMAND_SET_IMAGES} \
+  --set autoscaling.enabled=true \
+  --set autoscaling.enableWithExistingKEDA=false \
+  "
+fi
+
+if [ -n "${SET_MAX_REPLICAS}" ]; then
+  HELM_COMMAND_SET_IMAGES="${HELM_COMMAND_SET_IMAGES} \
+  --set autoscaling.scaledOptions.maxReplicaCount=${SET_MAX_REPLICAS} \
+  "
+fi
 
 if [ "${CHART_ENABLE_INGRESS_HOSTNAME}" = "true" ]; then
   if [[ ! $(cat /etc/hosts) == *"${HOSTNAME_ADDRESS}"* ]]; then
@@ -92,6 +127,15 @@ else
   HELM_COMMAND_SET_IMAGES="${HELM_COMMAND_SET_IMAGES} \
   --set global.K8S_PUBLIC_IP=${SELENIUM_GRID_HOST} \
   "
+  if [[ ! $(cat /etc/hosts) == *"alertmanager.selenium-grid.prod"* ]]; then
+    sudo -- sh -c -e "echo \"$(hostname -i) alertmanager.selenium-grid.prod\" >> /etc/hosts"
+  fi
+  if [[ ! $(cat /etc/hosts) == *"grafana.selenium-grid.prod"* ]]; then
+    sudo -- sh -c -e "echo \"$(hostname -i) grafana.selenium-grid.prod\" >> /etc/hosts"
+  fi
+  if [[ ! $(cat /etc/hosts) == *"pts.selenium-grid.prod"* ]]; then
+    sudo -- sh -c -e "echo \"$(hostname -i) pts.selenium-grid.prod\" >> /etc/hosts"
+  fi
 fi
 
 if [ "${CHART_ENABLE_BASIC_AUTH}" = "true" ]; then
@@ -116,6 +160,16 @@ HELM_COMMAND_SET_BASE_VALUES=" \
 --values ${RECORDER_VALUES_FILE} \
 --values ${TEST_VALUES_PATH}/base-resources-values.yaml \
 "
+
+if [ "${SUB_PATH}" = "/selenium" ]; then
+  HELM_COMMAND_SET_BASE_VALUES="${HELM_COMMAND_SET_BASE_VALUES} \
+  --values ${TEST_VALUES_PATH}/base-subPath-values.yaml \
+  "
+fi
+
+if [ "${SUB_PATH}" = "/" ]; then
+  SUB_PATH=""
+fi
 
 if [ "${SELENIUM_GRID_PROTOCOL}" = "https" ]; then
   HELM_COMMAND_SET_BASE_VALUES="${HELM_COMMAND_SET_BASE_VALUES} \