+++ tests/charts/templates/test.py
@@ -17,14 +17,14 @@ def load_template(yaml_file):
 
 class ChartTemplateTests(unittest.TestCase):
     def test_set_affinity(self):
-        resources_name = ['{0}-selenium-chrome-node'.format(RELEASE_NAME),
-                          '{0}-selenium-distributor'.format(RELEASE_NAME),
-                          '{0}-selenium-edge-node'.format(RELEASE_NAME),
-                          '{0}-selenium-firefox-node'.format(RELEASE_NAME),
-                          '{0}-selenium-event-bus'.format(RELEASE_NAME),
-                          '{0}-selenium-router'.format(RELEASE_NAME),
-                          '{0}-selenium-session-map'.format(RELEASE_NAME),
-                          '{0}-selenium-session-queue'.format(RELEASE_NAME)]
+        resources_name = ['{0}selenium-chrome-node'.format(RELEASE_NAME),
+                          '{0}selenium-distributor'.format(RELEASE_NAME),
+                          '{0}selenium-edge-node'.format(RELEASE_NAME),
+                          '{0}selenium-firefox-node'.format(RELEASE_NAME),
+                          '{0}selenium-event-bus'.format(RELEASE_NAME),
+                          '{0}selenium-router'.format(RELEASE_NAME),
+                          '{0}selenium-session-map'.format(RELEASE_NAME),
+                          '{0}selenium-session-queue'.format(RELEASE_NAME)]
         count = 0
         logger.info(f"Assert affinity is set in global and nodes")
         for doc in LIST_OF_DOCUMENTS:
@@ -35,7 +35,7 @@ class ChartTemplateTests(unittest.TestCase):
         self.assertEqual(count, len(resources_name), "Not all resources have affinity set")
 
     def test_ingress_nginx_annotations(self):
-        resources_name = ['{0}-selenium-ingress'.format(RELEASE_NAME)]
+        resources_name = ['{0}selenium-ingress'.format(RELEASE_NAME)]
         count = 0
         for doc in LIST_OF_DOCUMENTS:
             if doc['metadata']['name'] in resources_name and doc['kind'] == 'Ingress':
@@ -52,7 +52,7 @@ class ChartTemplateTests(unittest.TestCase):
         self.assertEqual(count, len(resources_name), "No ingress resources found")
 
     def test_sub_path_append_to_node_grid_url(self):
-        resources_name = ['{0}-selenium-node-config'.format(RELEASE_NAME)]
+        resources_name = ['{0}selenium-node-config'.format(RELEASE_NAME)]
         count = 0
         for doc in LIST_OF_DOCUMENTS:
             if doc['metadata']['name'] in resources_name and doc['kind'] == 'ConfigMap':
@@ -62,7 +62,7 @@ class ChartTemplateTests(unittest.TestCase):
         self.assertEqual(count, len(resources_name), "No node config resources found")
 
     def test_sub_path_set_to_grid_env_var(self):
-        resources_name = ['{0}-selenium-router'.format(RELEASE_NAME)]
+        resources_name = ['{0}selenium-router'.format(RELEASE_NAME)]
         is_present = False
         for doc in LIST_OF_DOCUMENTS:
             if doc['metadata']['name'] in resources_name and doc['kind'] == 'Deployment':
@@ -74,7 +74,7 @@ class ChartTemplateTests(unittest.TestCase):
         self.assertTrue(is_present, "ENV variable SE_SUB_PATH is not populated")
 
     def test_disable_ui_set_to_grid_env_var(self):
-        resources_name = ['{0}-selenium-router'.format(RELEASE_NAME)]
+        resources_name = ['{0}selenium-router'.format(RELEASE_NAME)]
         is_present = False
         for doc in LIST_OF_DOCUMENTS:
             if doc['metadata']['name'] in resources_name and doc['kind'] == 'Deployment':
@@ -86,18 +86,18 @@ class ChartTemplateTests(unittest.TestCase):
         self.assertTrue(is_present, "ENV variable SE_DISABLE_UI is not populated")
 
     def test_log_level_set_to_logging_config_map(self):
-        resources_name = ['{0}-selenium-chrome-node'.format(RELEASE_NAME),
-                          '{0}-selenium-distributor'.format(RELEASE_NAME),
-                          '{0}-selenium-edge-node'.format(RELEASE_NAME),
-                          '{0}-selenium-firefox-node'.format(RELEASE_NAME),
-                          '{0}-selenium-event-bus'.format(RELEASE_NAME),
-                          '{0}-selenium-router'.format(RELEASE_NAME),
-                          '{0}-selenium-session-map'.format(RELEASE_NAME),
-                          '{0}-selenium-session-queue'.format(RELEASE_NAME)]
+        resources_name = ['{0}selenium-chrome-node'.format(RELEASE_NAME),
+                          '{0}selenium-distributor'.format(RELEASE_NAME),
+                          '{0}selenium-edge-node'.format(RELEASE_NAME),
+                          '{0}selenium-firefox-node'.format(RELEASE_NAME),
+                          '{0}selenium-event-bus'.format(RELEASE_NAME),
+                          '{0}selenium-router'.format(RELEASE_NAME),
+                          '{0}selenium-session-map'.format(RELEASE_NAME),
+                          '{0}selenium-session-queue'.format(RELEASE_NAME)]
         logger.info(f"Assert log level value is set to logging ConfigMap")
         count_config = 0
         for doc in LIST_OF_DOCUMENTS:
-            if doc['metadata']['name'] == '{0}-selenium-logging-config'.format(RELEASE_NAME) and doc['kind'] == 'ConfigMap':
+            if doc['metadata']['name'] == '{0}selenium-logging-config'.format(RELEASE_NAME) and doc['kind'] == 'ConfigMap':
                 self.assertTrue(doc['data']['SE_LOG_LEVEL'] == 'FINE')
                 count_config += 1
         self.assertEqual(count_config, 1, "No logging ConfigMap found")
@@ -109,16 +109,16 @@ class ChartTemplateTests(unittest.TestCase):
                 list_env_from = doc['spec']['template']['spec']['containers'][0]['envFrom']
                 for env in list_env_from:
                     if env.get('configMapRef') is not None:
-                        if env['configMapRef']['name'] == '{0}-selenium-logging-config'.format(RELEASE_NAME):
+                        if env['configMapRef']['name'] == '{0}selenium-logging-config'.format(RELEASE_NAME):
                             is_present = True
                 self.assertTrue(is_present, "envFrom doesn't contain logging ConfigMap")
                 count += 1
         self.assertEqual(count, len(resources_name), "Logging ConfigMap is not present in expected resources")
 
     def test_node_port_set_when_service_type_is_node_port(self):
-        single_node_port = {'{0}-selenium-distributor'.format(RELEASE_NAME): 30553,
-                            '{0}-selenium-router'.format(RELEASE_NAME): 30444,
-                            '{0}-selenium-session-queue'.format(RELEASE_NAME): 30559}
+        single_node_port = {'{0}selenium-distributor'.format(RELEASE_NAME): 30553,
+                            '{0}selenium-router'.format(RELEASE_NAME): 30444,
+                            '{0}selenium-session-queue'.format(RELEASE_NAME): 30559}
         count = 0
         logger.info(f"Assert NodePort is set to components service")
         for doc in LIST_OF_DOCUMENTS:
@@ -130,16 +130,76 @@ class ChartTemplateTests(unittest.TestCase):
 
     def test_all_metadata_name_is_prefixed_with_release_name(self):
         logger.info(f"Assert all metadata name is prefixed with RELEASE NAME")
+        prefix = "selenium-" if RELEASE_NAME == "" else RELEASE_NAME
         for doc in LIST_OF_DOCUMENTS:
             logger.info(f"Assert metadata name: {doc['metadata']['name']}")
             self.assertTrue(doc['metadata']['name'].startswith(RELEASE_NAME),
                             f"Metadata name {doc['metadata']['name']} is not prefixed with RELEASE NAME: {RELEASE_NAME}")
+
+    def test_extra_script_import_to_node_configmap(self):
+        resources_name = ['{0}selenium-node-config'.format(RELEASE_NAME)]
+        count = 0
+        for doc in LIST_OF_DOCUMENTS:
+            if doc['metadata']['name'] in resources_name and doc['kind'] == 'ConfigMap':
+                logger.info(f"Assert default file is imported to Node ConfigMap")
+                self.assertTrue(doc['data']['nodeProbe.sh'] is not None)
+                self.assertTrue(doc['data']['nodePreStop.sh'] is not None)
+                self.assertTrue(doc['data']['nodeCustomTask.sh'] is not None)
+                self.assertTrue(doc['data']['setFromCommand.sh'] is not None)
+                count += 1
+        self.assertEqual(count, len(resources_name), "No node config resources found")
+
+    def test_extra_script_import_to_uploader_configmap(self):
+        resources_name = ['{0}selenium-uploader-config'.format(RELEASE_NAME)]
+        count = 0
+        for doc in LIST_OF_DOCUMENTS:
+            if doc['metadata']['name'] in resources_name and doc['kind'] == 'ConfigMap':
+                logger.info(f"Assert extra script is imported to Uploader ConfigMap")
+                self.assertTrue(doc['data']['upload.sh'] is not None)
+                self.assertTrue(doc['data']['setFromCommand.sh'] is not None)
+                count += 1
+        self.assertEqual(count, len(resources_name), "No uploader config resources found")
+
+    def test_extra_script_import_to_recorder_configmap(self):
+        resources_name = ['{0}selenium-recorder-config'.format(RELEASE_NAME)]
+        count = 0
+        for doc in LIST_OF_DOCUMENTS:
+            if doc['metadata']['name'] in resources_name and doc['kind'] == 'ConfigMap':
+                logger.info(f"Assert extra script is imported to Recorder ConfigMap")
+                self.assertTrue(doc['data']['video.sh'] is not None)
+                self.assertTrue(doc['data']['setFromCommand.sh'] is not None)
+                count += 1
+        self.assertEqual(count, len(resources_name), "No recorder config resources found")
+
+    def test_terminationGracePeriodSeconds_in_deployment_autoscaling(self):
+        resources_name = ['{0}selenium-chrome-node'.format(RELEASE_NAME),]
+        count = 0
+        for doc in LIST_OF_DOCUMENTS:
+            if doc['metadata']['name'] in resources_name and doc['kind'] == 'Deployment':
+                logger.info(f"Assert terminationGracePeriodSeconds is set in resource {doc['metadata']['name']}")
+                self.assertTrue(doc['spec']['template']['spec']['terminationGracePeriodSeconds'] == 7200)
+                count += 1
+        self.assertEqual(count, len(resources_name), "node.terminationGracePeriodSeconds doesn't override a higher value than autoscaling.terminationGracePeriodSeconds")
+
+        resources_name = ['{0}selenium-edge-node'.format(RELEASE_NAME),
+                          '{0}selenium-firefox-node'.format(RELEASE_NAME),]
+        count = 0
+        for doc in LIST_OF_DOCUMENTS:
+            if doc['metadata']['name'] in resources_name and doc['kind'] == 'Deployment':
+                logger.info(f"Assert terminationGracePeriodSeconds is set in resource {doc['metadata']['name']}")
+                self.assertTrue(doc['spec']['template']['spec']['terminationGracePeriodSeconds'] == 3600)
+                count += 1
+        self.assertEqual(count, len(resources_name), "node.terminationGracePeriodSeconds doesn't inherit the global value autoscaling.terminationGracePeriodSeconds")
 
 if __name__ == '__main__':
     failed = False
     try:
         FILE_NAME = sys.argv[1]
         RELEASE_NAME = sys.argv[2]
+        if RELEASE_NAME == "selenium":
+            RELEASE_NAME = ""
+        else:
+            RELEASE_NAME = RELEASE_NAME + "-"
         LIST_OF_DOCUMENTS = load_template(FILE_NAME)
         suite = unittest.TestLoader().loadTestsFromTestCase(ChartTemplateTests)
         test_runner = unittest.TextTestRunner(verbosity=3)