+++ charts/selenium-grid/CHANGELOG.md
@@ -1,3 +1,63 @@
+## :heavy_check_mark: selenium-grid-0.28.4
+
+- Chart is using image tag 4.18.1-20240224
+- Chart is tested on Kubernetes versions: v1.25.16 v1.26.14 v1.27.11 v1.28.7 v1.29.2 
+- Chart is tested on Helm versions: v3.10.3 v3.11.3 v3.12.3 v3.13.3 v3.14.2 
+
+### Fixed
+- fix(chart): connection in script of Node startup probe and preStop lifecycle :: Viet Nguyen Duc
+- fix(chart): `autoscaling.terminationGracePeriodSeconds` is not set in Node spec :: Viet Nguyen Duc
+
+### Changed
+- Release chart 0.28.4 :: Viet Nguyen Duc
+- update(chart): add annotations checksum for ConfigMap and Secret :: Viet Nguyen Duc
+- test(chart): update chart values for CI tests :: Viet Nguyen Duc
+- test(chart): autoscaling as job :: Viet Nguyen Duc
+- Update tag in docs and files :: Selenium CI Bot
+
+## :heavy_check_mark: selenium-grid-0.28.3
+
+- Chart is using image tag 4.18.1-20240224
+- Chart is tested on Kubernetes versions: v1.25.16 v1.26.14 v1.27.11 v1.28.7 v1.29.2 
+
+### Fixed
+- fix(chart): fix object naming and add test to verify :: Viet Nguyen Duc
+
+### Changed
+- Update tag in docs and files :: Selenium CI Bot
+
+## :heavy_check_mark: selenium-grid-0.28.2
+
+- Chart is using image tag 4.18.1-20240224
+- Chart is tested on Kubernetes versions: v1.25.16 v1.26.14 v1.27.11 v1.28.7 v1.29.2 
+
+### Added
+- feat(chart): option to disable release name prefix in object naming :: Viet Nguyen Duc
+
+### Fixed
+- fix(chart): extra scripts can be imported in sub-chart by default :: Viet Nguyen Duc
+
+### Changed
+- release(chart): 0.28.2 :: Viet Nguyen Duc
+- Update tag in docs and files :: Selenium CI Bot
+- test(chart): test extra scripts import when import as sub-chart :: Viet Nguyen Duc
+
+## :heavy_check_mark: selenium-grid-0.28.1
+
+- Chart is using image tag 4.18.0-20240220
+- Chart is tested on Kubernetes versions: v1.25.16 v1.26.14 v1.27.11 v1.28.7 v1.29.2
+
+### Fixed
+- bug(chart): template issue when chart is imported as dependency in umbrella charts :: Viet Nguyen Duc
+- bug(chart): SE_NODE_GRID_URL missing port when `hostname` is `selenium-grid.local` :: Viet Nguyen Duc
+- bug(chart) CRITICAL: Node startup probe loop infinite when ingress hostname is set :: Viet Nguyen Duc
+
+### Changed
+- test(chart): update docs :: Viet Nguyen Duc
+- test(chart): add tests for the case basic auth is enabled :: Viet Nguyen Duc
+- test(chart): add tests for the case ingress is enabled with `hostname` set :: Viet Nguyen Duc
+- build(chart): change log and release notes for helm chart :: Viet Nguyen Duc
+
 ## :heavy_check_mark: selenium-grid-0.28.0
 
 - Chart is using image tag 4.18.0-20240220